Criação dos logs de fraudes.
- Criação do log de fraudes.

+++ zenon-fraud-detector/src/main/java/br/com/zenon/fraud/Customer.java
@@ -1,5 +1,7 @@
 package br.com.zenon.fraud;
 
-public record Customer(String name, double oldBalance, double newBalance) {
+import java.math.BigDecimal;
+
+public record Customer(String name, BigDecimal oldBalance, BigDecimal newBalance) {
 
 }
